Hardware tabs drop ?at= once the reader moves the window
Network Performance and Device Stats already do this: a link's at= says where it wanted to
put you, and left on the URL after you move, a reload drags you back to it. The five hardware
stat tabs kept theirs, so refreshing after changing the window returned to the alert's moment.

They report through the ref and JSInvokable the analysis charts already use, so the drop
happens in the one navigation ClearLinkedInstant was written to make, rather than a second
replaceState behind Blazor's router. Reported from every gesture that moves the window: preset,
shift, custom apply, drag zoom.
